+++ Participantes/src/main/java/org/acme/Controller/AlunoController.java
@@ -26,6 +26,8 @@ public class AlunoController {
     @GET
     @Path("/{id}")
     public Response findById(@PathParam("id") Long id) {
+        if(service.findById(id) == null){return Response.status(Response.Status.NOT_FOUND).build();}
+
         return Response.ok(service
                 .findById(id))
                 .build();
@@ -42,6 +44,7 @@ public class AlunoController {
     @PUT
     @Path("/{id}")
     public Response update(@PathParam("id") Long id, Aluno Aluno) {
+        if(service.findById(id) == null){return Response.status(Response.Status.NOT_FOUND).build();}
         return Response.ok(service
                 .update(id, Aluno))
                 .build();
@@ -50,6 +53,7 @@ public class AlunoController {
     @DELETE
     @Path("/{id}")
     public Response delete(@PathParam("id") Long id) {
+        if(service.findById(id) == null){return Response.status(Response.Status.NOT_FOUND).build();}
         service.delete(id);
         return Response
                 .noContent()
